refactor: modularize and encapsulate logic into files & functions
Key changes:
- Selecting something is possible through a custom `choice` function
- Logic is seperated into multiple files for ease of use
