+++ go.mod
@@ -2,9 +2,10 @@ module github.com/intercepted16/CleanDL
 
 go 1.18
 
+require github.com/urfave/cli/v2 v2.27.2
+
 require (
 	github.com/cpuguy83/go-md2man/v2 v2.0.4 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
-	github.com/urfave/cli/v2 v2.27.2 // indirect
 	github.com/xrash/smetrics v0.0.0-20240312152122-5f08fbb34913 // indirect
 )
